modularisation
URIParser.h no longer has lines that print out, this is to ensure that when the class is used in other projects, the users can control what to print to the screen (or UI). A few classes added to allow printing of proper outputs as well as errors.

Try-catch block fixed in main.cpp to add some error handling for the UI only, the class should not need any error handling.

Also fixed the regex and string checks in the class so now the scheme, user&password and host are correct, with the other 2 URI types (file and URN) being "completed".

Only missing parts are port, path (for the URL-URI), query and fragment i.e. only need to finish the URL parser