testing
tested the current parts of the program to ensure they are not breaking, by introducing different types of URIs and they seem to work, except for the domain regex check.

need to do, path, query and fragment to finish this task.

optimised quite a bit of the functions and removed a lot of if statement to streamline the flow of the code

also separated the parsing functions from the URI type check function into 3 functions that parse each type of URI